Crate cubecl_core

Re-exports

pub use frontend::cmma;

pub use prelude::CubeDim;

pub use prelude::flex32;

pub use prelude::tf32;

pub use codegen::*;

Modules

benchmark

Module for benchmarking any executable part

channel

Compute channel module.

client

Compute client module.

codegen

compute

frontend

Cube Frontend Types.

future

Some future utilities that work across environments. Future utils with a compatible API for native, non-std and wasm environments.

ir

Cube Language Internal Representation.

prelude

server

Compute server module.

tune

Autotune module

Macros

comment

Insert a literal comment into the kernel source code.

comptime

Mark the contents of this macro as compile time values, turning off all expansion for this code and using it verbatim

cpa

Cube Pseudo Assembly.

debug_print

Print a formatted message using the target’s debug print facilities. The format string is target specific, but Vulkan and CUDA both use the C++ conventions. WGSL isn’t currently supported.

debug_print_expand

Print a formatted message using the target’s debug print facilities. The format string is target specific, but Vulkan and CUDA both use the C++ conventions. WGSL isn’t currently supported.

unexpanded

Structs

KernelId

Kernel unique identifier.

Enums

AtomicFeature

CubeCount

Specifieds the number of cubes to be dispatched for a kernel.

ExecutionMode

The kind of execution to be performed.

Feature

Every feature that can be supported by a cube runtime.

MemoryConfiguration

High level configuration of memory management.

Constants

PLANE_DIM_APPROX

An approximation of the plane dimension.

Traits

CubeElement

The base element trait for the jit backend.

Kernel

Implement this trait to create a kernel definition.

Runtime

Runtime for the CubeCL.

Functions

calculate_cube_count_elemwise

Calculate the number of cubes required to execute an operation where one cube unit is assigned to one element.

tensor_line_size

tensor_line_size_parallel

Find the maximum line size usable for parallel vectorization along the given axis from the supported line sizes or return 1 if vectorization is impossible.

tensor_line_size_perpendicular

Find the maximum line size usable for perpendicular vectorization along the given axis from the supported line sizes or return 1 if vectorization is impossible.

tensor_vectorization_factor

Type Aliases

RuntimeArg

Runtime arguments to launch a kernel.

Attribute Macros

cube

Mark a cube function, trait or implementation for expansion.

tune

Crates a tuning set with a specific signature. Should return a tuple of benchmark inputs.

Derive Macros

AutotuneKey

Implements display and initialization for autotune keys.

CubeLaunch

Derive macro to define a cube type that is launched with a kernel

CubeType

Derive macro to define a cube type that is not launched