cubecl_core::compute

Trait CubeTask

pub trait CubeTask<C: Compiler>: Send + Sync {
    // Required methods
    fn id(&self) -> KernelId;
    fn compile(
        &self,
        compilation_options: &C::CompilationOptions,
        mode: ExecutionMode,
    ) -> CompiledKernel<C>;

    // Provided method
    fn name(&self) -> &'static str { ... }
}

Kernel trait with the ComputeShader that will be compiled and cached based on the provided id.

Required Methods

fn id(&self) -> KernelId

Identifier for the kernel, used for caching kernel compilation.

fn compile( &self, compilation_options: &C::CompilationOptions, mode: ExecutionMode, ) -> CompiledKernel<C>

Compile the kernel into source

Provided Methods

fn name(&self) -> &'static str