cubecl_core::frontend

Struct SharedMemory

pub struct SharedMemory<T: CubeType> { /* private fields */ }

Implementations

impl<E: CubePrimitive> SharedMemory<E>

pub unsafe fn index_unchecked<I: Index>(&self, _i: I) -> &E
where Self: CubeIndex<I>,

Perform an unchecked index into the array

Safety

Out of bounds indexing causes undefined behaviour and may segfault. Ensure index is always in bounds

pub unsafe fn index_assign_unchecked<I: Index>(&mut self, _i: I, _value: E)
where Self: CubeIndexMut<I>,

Perform an unchecked index assignment into the array

Safety

Out of bounds indexing causes undefined behaviour and may segfault. Ensure index is always in bounds

impl<T: CubePrimitive + Clone> SharedMemory<T>

pub fn new<S: Index>(_size: S) -> Self

pub fn new_lined<S: Index>( _size: S, _vectorization_factor: u32, ) -> SharedMemory<Line<T>>

pub fn __expand_new_lined( context: &mut CubeContext, size: ExpandElementTyped<u32>, vectorization_factor: u32, ) -> <SharedMemory<Line<T>> as CubeType>::ExpandType

pub fn vectorized<S: Index>(_size: S, _vectorization_factor: u32) -> Self

pub fn __expand_vectorized( context: &mut CubeContext, size: ExpandElementTyped<u32>, vectorization_factor: u32, ) -> <Self as CubeType>::ExpandType

pub fn __expand_new( context: &mut CubeContext, size: ExpandElementTyped<u32>, ) -> <Self as CubeType>::ExpandType