cubecl_core::frontend

Trait BitCast

pub trait BitCast: CubePrimitive {
    // Provided methods
    fn bitcast_from<From: CubePrimitive>(value: From) -> Self { ... }
    fn __expand_bitcast_from<From: CubePrimitive>(
        context: &mut CubeContext,
        value: ExpandElementTyped<From>,
    ) -> <Self as CubeType>::ExpandType { ... }
}

Enables reinterpet-casting/bitcasting from any floating point value to any integer value and vice versa

Provided Methods

fn bitcast_from<From: CubePrimitive>(value: From) -> Self

Reinterpret the bits of another primitive as this primitive without conversion.

fn __expand_bitcast_from<From: CubePrimitive>( context: &mut CubeContext, value: ExpandElementTyped<From>, ) -> <Self as CubeType>::ExpandType

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.