cubecl_core

Module ir

Cube Language Internal Representation.

Re-exports

pub use super::frontend::AtomicOp;

Structs

Allocator

An allocator for local variables of a kernel.

BinaryOperator

Binding

ClampOperator

CompareAndSwapOperator

CopyMemoryBulkOperator

CopyMemoryOperator

CubeDim

FmaOperator

If

IfElse

Instruction

An instruction that contains a right hand side Operation and an optional out variable.

Item

KernelDefinition

LineInitOperator

Loop

Matrix

RangeLoop

ReadGlobalOperator

ReadGlobalWithLayoutOperator

Scope

The scope is the main operation and variable container that simplify the process of reading inputs, creating local variables and adding new operations.

Select

SliceOperator

Switch

UnaryOperator

UnrolledRangeLoop

Variable

Enums

Branch

All branching types.

Builtin

ConstantScalarValue

The scalars are stored with the highest precision possible, but they might get reduced during compilation.

CoopMma

Cooperative Matrix-Multiply and Accumulate Instruction.

Elem

FloatKind

IntKind

Location

MatrixIdent

MatrixLayout

Metadata

All metadata that can be accessed in a shader.

NonSemantic

Operations that don’t change the semantics of the kernel. In other words, operations that do not perform any computation, if they run at all. i.e. println, comments and debug symbols.

Operation

All operations that can be used in a GPU compute shader.

Operator

All operators that can be used in a GPU compute shader.

Plane

All plane operations.

ReadingStrategy

Synchronization

All synchronization types.

UIntKind

VariableKind

Visibility

Functions

expand_checked_index_assign

fmt_vararg

Type Aliases

Id

Vectorization