cubecl_core::ir

Enum Elem

pub enum Elem {
    Float(FloatKind),
    Int(IntKind),
    UInt(UIntKind),
    AtomicFloat(FloatKind),
    AtomicInt(IntKind),
    AtomicUInt(UIntKind),
    Bool,
}

Variants

Float(FloatKind)

Int(IntKind)

UInt(UIntKind)

AtomicFloat(FloatKind)

AtomicInt(IntKind)

AtomicUInt(UIntKind)

Bool

Implementations

impl Elem

pub fn constant_from_f64(&self, val: f64) -> Variable

Create a constant scalar from a float.

The output will have the same type as the element.

pub fn constant_from_i64(&self, val: i64) -> Variable

Create a constant scalar from a signed integer.

The output will have the same type as the element.

pub fn constant_from_u64(&self, val: u64) -> Variable

Create a constant scalar from a unsigned integer.

The output will have the same type as the element.

pub fn constant_from_bool(&self, val: bool) -> Variable

Create a constant scalar from a boolean.

The output will have the same type as the element.

pub fn from_constant(&self, constant: Variable) -> Variable

Ensure that the variable provided, when a constant, is the same type as elem.

pub const fn size(&self) -> usize

Get the size in bytes.

pub fn is_atomic(&self) -> bool

pub fn is_int(&self) -> bool

pub fn max_variable(&self) -> Variable

pub fn min_variable(&self) -> Variable