cubecl_core

Module prelude

Re-exports

pub use crate::Kernel;

pub use crate::RuntimeArg;

pub use crate::codegen::KernelExpansion;

pub use crate::codegen::KernelIntegrator;

pub use crate::codegen::KernelSettings;

pub use crate::compute::CompiledKernel;

pub use crate::compute::CubeTask;

pub use crate::compute::KernelBuilder;

pub use crate::compute::KernelLauncher;

pub use crate::compute::KernelTask;

pub use crate::frontend::cmma;

pub use crate::ir::CubeDim;

pub use crate::ir::KernelDefinition;

pub use crate::frontend::Array;

pub use crate::frontend::ArrayHandleRef;

pub use crate::frontend::Atomic;

pub use crate::frontend::Float;

pub use crate::frontend::FloatExpand;

pub use crate::frontend::LaunchArg;

pub use crate::frontend::NumericExpand;

pub use crate::frontend::Slice;

pub use crate::frontend::SliceMut;

pub use crate::frontend::Tensor;

pub use crate::frontend::TensorArg;

pub use crate::frontend::plane_all;

pub use crate::frontend::plane_all;

pub use crate::frontend::plane_max;

pub use crate::frontend::plane_max;

pub use crate::frontend::plane_min;

pub use crate::frontend::plane_min;

pub use crate::frontend::plane_prod;

pub use crate::frontend::plane_prod;

pub use crate::frontend::plane_sum;

pub use crate::frontend::plane_sum;

pub use crate::frontend::branch::*;

pub use crate::frontend::synchronization::*;

pub use crate::frontend::*;

Macros

comment

Insert a literal comment into the kernel source code.

comptime

Mark the contents of this macro as compile time values, turning off all expansion for this code and using it verbatim

Structs

ComputeClient

The ComputeClient is the entry point to require tasks from the ComputeServer. It should be obtained for a specific device via the Compute struct.

Enums

CubeCount

Specifieds the number of cubes to be dispatched for a kernel.

Traits

CubeElement

The base element trait for the jit backend.

Runtime

Runtime for the CubeCL.

Attribute Macros

cube

Mark a cube function, trait or implementation for expansion.

Derive Macros

CubeLaunch

Derive macro to define a cube type that is launched with a kernel

CubeType

Derive macro to define a cube type that is not launched