cubecl_core

Trait CubeElement

pub trait CubeElement:
    Debug
    + Send
    + Sync
    + 'static
    + Clone
    + Pod {
    // Required methods
    fn type_name() -> &'static str;
    fn as_bytes(slice: &[Self]) -> &[u8] ;
    fn from_bytes(bytes: &[u8]) -> &[Self];
    fn cube_elem() -> Elem;
    fn maximum_value() -> Self;
    fn minimum_value() -> Self;
}

The base element trait for the jit backend.

Required Methods

fn type_name() -> &'static str

Returns the name of the type.

fn as_bytes(slice: &[Self]) -> &[u8]

Convert a slice of elements to a slice of bytes.

fn from_bytes(bytes: &[u8]) -> &[Self]

Convert a slice of bytes to a slice of elements.

fn cube_elem() -> Elem

Element representation for cubecl.

fn maximum_value() -> Self

Highest possible value

fn minimum_value() -> Self

Lowest possible value

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.