cubecl_core

Trait Runtime

pub trait Runtime:
    Send
    + Sync
    + 'static
    + Debug {
    type Compiler: Compiler;
    type Server: ComputeServer<Kernel = Box<dyn CubeTask<Self::Compiler>>, Feature = Feature>;
    type Channel: ComputeChannel<Self::Server>;
    type Device: Default + Clone + Debug;

    // Required methods
    fn client(
        device: &Self::Device,
    ) -> ComputeClient<Self::Server, Self::Channel>;
    fn name() -> &'static str;
    fn extension() -> &'static str;
    fn supported_line_sizes() -> &'static [u8] ;
    fn max_cube_count() -> (u32, u32, u32);

    // Provided methods
    fn require_array_lengths() -> bool { ... }
    fn line_size_elem(elem: &Elem) -> impl Iterator<Item = u8> + Clone { ... }
}

Runtime for the CubeCL.

Required Associated Types

type Compiler: Compiler

The compiler used to compile the inner representation into tokens.

type Server: ComputeServer<Kernel = Box<dyn CubeTask<Self::Compiler>>, Feature = Feature>

The compute server used to run kernels and perform autotuning.

type Channel: ComputeChannel<Self::Server>

The channel used to communicate with the compute server.

type Device: Default + Clone + Debug

The device used to retrieve the compute client.

Required Methods

fn client(device: &Self::Device) -> ComputeClient<Self::Server, Self::Channel>

Retrieve the compute client from the runtime device.

fn name() -> &'static str

The runtime name.

fn extension() -> &'static str

The default extension for the runtime’s kernel/shader code. Might change based on which compiler is used.

fn supported_line_sizes() -> &'static [u8]

Returns the supported line sizes for the current runtime’s compiler.

fn max_cube_count() -> (u32, u32, u32)

Returns the maximum cube count on each dimension that can be launched.

Provided Methods

fn require_array_lengths() -> bool

Return true if global input array lengths should be added to kernel info.

fn line_size_elem(elem: &Elem) -> impl Iterator<Item = u8> + Clone

Returns all line sizes that are useful to perform IO operation on the given element.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.